separate base class
